\name{tuning}
\alias{tuning}
\alias{tuning<-}
\title{Tuning parameters for Metropolis-Hastings proposals}
\description{
  Accessor and replacement methods for tuning the Metropolis-Hastings
  proposal parameters.
}
\usage{
tuning(object)
}
\arguments{
  \item{object}{ Object of class \code{XdeParameter} }
}
\details{
  See the XdeParameterClass vignette
}
\value{
  A numerical vector
}
\author{R. Scharpf}
\keyword{methods}

